code/Web/backend/requirements.txt
fastapi==0.115.11
mysql-connector-python==9.2.0
mysql_connector_repackaged==0.3.1
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.27.0